requirement.txt
Django==1.7.1
South==1.0.2
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
celery==3.1.17
django-filter==0.9.1
djangorestframework==3.0.2
ipython==2.3.1
kombu==3.0.24
psycopg2==2.5.4
pytz==2014.10
wsgiref==0.1.2
